Fixed issues presented by code reviewer
